Modify opsmgr charm to use shared-db interface
The shared-db interface is necessary in an HA configuration to ensure
that multiple instances of the same charm share the same database.

Using the shared-db interface also helps with the problem of using
multiple networks, in relationship-joined you tell the database
what ip you want to connect to it using, and the database returns
it's own ip for that same network.

Change-Id: If1dcbccef81a20f9c1ffdfce3a4a56818975fc25